feat: API 변경에 따른 전체 스키장 정보 API 적용 (홈화면)

+++ app/src/main/java/com/dieski/weski/navigation/MainNavigator.kt
@@ -7,7 +7,6 @@ import androidx.navigation.NavDestination.Companion.hasRoute
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.dieski.weski.presentation.core.model.WeatherType
 import com.dieski.weski.presentation.core.navigation.Route
 import com.dieski.weski.presentation.detail.navigation.navigateDetail
 
@@ -42,7 +41,7 @@ internal class MainNavigator(
     }*/
 
     fun navigateToDetail(
-        resortId: Int,
+        resortId: Long,
         resortName: String,
         resortWebKey: String,
         temperature: Int,
