+++ presentation/src/main/java/com/dieski/weski/presentation/core/navigation/RouteModel.kt
@@ -9,7 +9,7 @@ sealed interface Route {
 
     @Serializable
     data class Detail(
-        val resortId: Int,
+        val resortId: Long,
         val resortName: String,
         val resortWebKey: String,
         val temperature: Int,
